+++ src/ensembl/src/content/app/browser/browser-nav/BrowserNavIcon.tsx
@@ -16,7 +16,7 @@
 
 import React, { FunctionComponent, memo } from 'react';
 
-import browserMessagingService from 'src/content/app/browser/browser-messaging-service';
+import browserMessagingService from 'src/content/app/browser/services/browser-messaging-service/browser-messaging-service';
 
 import ImageButton from 'src/shared/components/image-button/ImageButton';
