+++ src/ensembl/src/content/app/browser/services/browser-messaging-service/browser-messaging-service.test.ts
@@ -20,11 +20,14 @@ import faker from 'faker';
 import windowService from 'src/services/window-service';
 
 class MockWindow {
-  private subscribers: { [name: string]: Function[] } = {};
+  private subscribers: { [name: string]: ((...args: any[]) => void)[] } = {};
 
   public postMessage = jest.fn();
 
-  public addEventListener = (name: string, callback: Function) => {
+  public addEventListener = (
+    name: string,
+    callback: (...args: any[]) => void
+  ) => {
     if (!this.subscribers[name]) {
       this.subscribers[name] = [];
     }
